+++ apps/snap/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.3.2
+
+### Patch Changes
+
+- use opts for get notes opts single for getBalanceForAsset
+
+## 0.3.1
+
+### Patch Changes
+
+- 14d0ac58: fix snap<>fe-sdk bug, ensure fe-sdk stringifies all snap params, ensure snap parses them
+
 ## 0.3.0
 
 ### Minor Changes